#294: Refactored class Column once again
