+++ exasol/analytics/schema/column_type_utils.py
